subscription: got preliminary data flow working
got the data flow working so that redis handles cache storage, publish /
subscribe, and task queues, so that the forms autosave and collaboration
is done over websockets handled by graphql :). there was a bug in the
vue-apollo subscribe library, so as a workaround I am not using
variables. hopefully this can be fixed eventually

Signed-off-by: Joshua <jschmidtnj@gmail.com>